More tweaks to day 17 part 1

+++ 2022/day-12/src/main/kotlin/com/closeratio/aoc2022/day12/HeightMapParser.kt
@@ -14,7 +14,7 @@ class HeightMapParser(
         .flatMapIndexed { y, line ->
             line.mapIndexed { x, char ->
                 Tile(
-                    Vec2(x.toLong(), y.toLong()),
+                    Vec2(x, y),
                     char.toString()
                 )
             }
