+++ 2022/day-23/src/main/kotlin/com/closeratio/aoc2022/day23/WorldStateSimulationParser.kt
@@ -18,7 +18,7 @@ class WorldStateSimulationParser(
         .flatMapIndexed { y, line ->
             line.mapIndexedNotNull { x, char ->
                 when (char) {
-                    '#' -> Vec2(x.toLong(), y.toLong())
+                    '#' -> Vec2(x, y)
                     else -> null
                 }
             }
