+++ 2023/day-11/src/main/kotlin/com/closeratio/aoc2023/day11/GalaxyParser.kt
@@ -12,7 +12,7 @@ class GalaxyParser {
         .flatMapIndexed { y, line ->
             line.mapIndexedNotNull { x, char ->
                 when (char) {
-                    '#' -> Vec2(x.toLong(), y.toLong())
+                    '#' -> Vec2(x, y)
                     else -> null
                 }
             }
