+++ 2023/day-13/src/main/kotlin/com/closeratio/aoc2023/day13/MirrorAnalyser.kt
@@ -13,7 +13,7 @@ class MirrorAnalyser {
         .flatMapIndexed { y, line ->
             line.mapIndexedNotNull { x, char ->
                 when (char) {
-                    '#' -> Vec2(x.toLong(), y.toLong())
+                    '#' -> Vec2(x, y)
                     else -> null
                 }
             }
