+++ 2023/day-3/src/main/kotlin/com/closeratio/aoc2023/day3/SchematicAnalyser.kt
@@ -9,7 +9,7 @@ class SchematicAnalyser {
     private fun parseSchematic(input: List<String>): EngineSchematic = input
         .flatMapIndexed { y, line ->
             line.mapIndexed { x, char ->
-                Vec2(x.toLong(), y.toLong()) to char.toString()
+                Vec2(x, y) to char.toString()
             }
         }
         .toMap()
